+++ include/crath/simd/float2x4.h
@@ -45,11 +45,11 @@ namespace cr::simd
 		      f2(_mm_set_ps(a7, a6, a5, a4)) {
 		}
 
-		inline float const& operator[](size_t i) const {
+		inline float const& operator[](int64_t i) const {
 			return const_cast<float2x4*>(this)->operator[](i);
 		}
 
-		inline float& operator[](size_t i) {
+		inline float& operator[](int64_t i) {
 			if (i < 4) {
 #if defined(COMPILER_MSVC)
 				return this->f1.m128_f32[i];
